chatbotknowledge base Task-1/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
feedparser
apscheduler
chromadb
openai
python-dotenv
tiktoken
pinecone-client
sentence-transformers
